# backend/api/userfeedback/v1/user_feedback_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v3.21.12
// source: userfeedback/v1/user_feedback.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationUserFeedbackCreateFeedback = "/userfeedback.v1.UserFeedback/CreateFeedback"
const OperationUserFeedbackDeleteFeedback = "/userfeedback.v1.UserFeedback/DeleteFeedback"
const OperationUserFeedbackGetFeedback = "/userfeedback.v1.UserFeedback/GetFeedback"
const OperationUserFeedbackListFeedbacks = "/userfeedback.v1.UserFeedback/ListFeedbacks"
const OperationUserFeedbackListFeedbacksByStatus = "/userfeedback.v1.UserFeedback/ListFeedbacksByStatus"
const OperationUserFeedbackUpdateFeedback = "/userfeedback.v1.UserFeedback/UpdateFeedback"

type UserFeedbackHTTPServer interface {
	// CreateFeedback Create a new user feedback
	CreateFeedback(context.Context, *CreateFeedbackRequest) (*CreateFeedbackReply, error)
	// DeleteFeedback Delete a user feedback
	DeleteFeedback(context.Context, *DeleteFeedbackRequest) (*DeleteFeedbackReply, error)
	// GetFeedback Get a user feedback by ID
	GetFeedback(context.Context, *GetFeedbackRequest) (*GetFeedbackReply, error)
	// ListFeedbacks List user feedbacks with pagination
	ListFeedbacks(context.Context, *ListFeedbacksRequest) (*ListFeedbacksReply, error)
	// ListFeedbacksByStatus List user feedbacks by status
	ListFeedbacksByStatus(context.Context, *ListFeedbacksByStatusRequest) (*ListFeedbacksByStatusReply, error)
	// UpdateFeedback Update a user feedback
	UpdateFeedback(context.Context, *UpdateFeedbackRequest) (*UpdateFeedbackReply, error)
}

func RegisterUserFeedbackHTTPServer(s *http.Server, srv UserFeedbackHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/userfeedback", _UserFeedback_CreateFeedback0_HTTP_Handler(srv))
	r.GET("/v1/userfeedback/{id}", _UserFeedback_GetFeedback0_HTTP_Handler(srv))
	r.GET("/v1/userfeedback", _UserFeedback_ListFeedbacks0_HTTP_Handler(srv))
	r.PUT("/v1/userfeedback/{id}", _UserFeedback_UpdateFeedback0_HTTP_Handler(srv))
	r.DELETE("/v1/userfeedback/{id}", _UserFeedback_DeleteFeedback0_HTTP_Handler(srv))
	r.GET("/v1/userfeedback/status/{status}", _UserFeedback_ListFeedbacksByStatus0_HTTP_Handler(srv))
}

func _UserFeedback_CreateFeedback0_HTTP_Handler(srv UserFeedbackHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateFeedbackRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserFeedbackCreateFeedback)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateFeedback(ctx, req.(*CreateFeedbackRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateFeedbackReply)
		return ctx.Result(200, reply)
	}
}

func _UserFeedback_GetFeedback0_HTTP_Handler(srv UserFeedbackHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetFeedbackRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserFeedbackGetFeedback)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetFeedback(ctx, req.(*GetFeedbackRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetFeedbackReply)
		return ctx.Result(200, reply)
	}
}

func _UserFeedback_ListFeedbacks0_HTTP_Handler(srv UserFeedbackHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListFeedbacksRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserFeedbackListFeedbacks)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListFeedbacks(ctx, req.(*ListFeedbacksRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListFeedbacksReply)
		return ctx.Result(200, reply)
	}
}

func _UserFeedback_UpdateFeedback0_HTTP_Handler(srv UserFeedbackHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateFeedbackRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserFeedbackUpdateFeedback)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateFeedback(ctx, req.(*UpdateFeedbackRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateFeedbackReply)
		return ctx.Result(200, reply)
	}
}

func _UserFeedback_DeleteFeedback0_HTTP_Handler(srv UserFeedbackHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteFeedbackRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserFeedbackDeleteFeedback)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteFeedback(ctx, req.(*DeleteFeedbackRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteFeedbackReply)
		return ctx.Result(200, reply)
	}
}

func _UserFeedback_ListFeedbacksByStatus0_HTTP_Handler(srv UserFeedbackHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListFeedbacksByStatusRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserFeedbackListFeedbacksByStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListFeedbacksByStatus(ctx, req.(*ListFeedbacksByStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListFeedbacksByStatusReply)
		return ctx.Result(200, reply)
	}
}

type UserFeedbackHTTPClient interface {
	// CreateFeedback Create a new user feedback
	CreateFeedback(ctx context.Context, req *CreateFeedbackRequest, opts ...http.CallOption) (rsp *CreateFeedbackReply, err error)
	// DeleteFeedback Delete a user feedback
	DeleteFeedback(ctx context.Context, req *DeleteFeedbackRequest, opts ...http.CallOption) (rsp *DeleteFeedbackReply, err error)
	// GetFeedback Get a user feedback by ID
	GetFeedback(ctx context.Context, req *GetFeedbackRequest, opts ...http.CallOption) (rsp *GetFeedbackReply, err error)
	// ListFeedbacks List user feedbacks with pagination
	ListFeedbacks(ctx context.Context, req *ListFeedbacksRequest, opts ...http.CallOption) (rsp *ListFeedbacksReply, err error)
	// ListFeedbacksByStatus List user feedbacks by status
	ListFeedbacksByStatus(ctx context.Context, req *ListFeedbacksByStatusRequest, opts ...http.CallOption) (rsp *ListFeedbacksByStatusReply, err error)
	// UpdateFeedback Update a user feedback
	UpdateFeedback(ctx context.Context, req *UpdateFeedbackRequest, opts ...http.CallOption) (rsp *UpdateFeedbackReply, err error)
}

type UserFeedbackHTTPClientImpl struct {
	cc *http.Client
}

func NewUserFeedbackHTTPClient(client *http.Client) UserFeedbackHTTPClient {
	return &UserFeedbackHTTPClientImpl{client}
}

// CreateFeedback Create a new user feedback
func (c *UserFeedbackHTTPClientImpl) CreateFeedback(ctx context.Context, in *CreateFeedbackRequest, opts ...http.CallOption) (*CreateFeedbackReply, error) {
	var out CreateFeedbackReply
	pattern := "/v1/userfeedback"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserFeedbackCreateFeedback))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// DeleteFeedback Delete a user feedback
func (c *UserFeedbackHTTPClientImpl) DeleteFeedback(ctx context.Context, in *DeleteFeedbackRequest, opts ...http.CallOption) (*DeleteFeedbackReply, error) {
	var out DeleteFeedbackReply
	pattern := "/v1/userfeedback/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserFeedbackDeleteFeedback))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetFeedback Get a user feedback by ID
func (c *UserFeedbackHTTPClientImpl) GetFeedback(ctx context.Context, in *GetFeedbackRequest, opts ...http.CallOption) (*GetFeedbackReply, error) {
	var out GetFeedbackReply
	pattern := "/v1/userfeedback/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserFeedbackGetFeedback))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListFeedbacks List user feedbacks with pagination
func (c *UserFeedbackHTTPClientImpl) ListFeedbacks(ctx context.Context, in *ListFeedbacksRequest, opts ...http.CallOption) (*ListFeedbacksReply, error) {
	var out ListFeedbacksReply
	pattern := "/v1/userfeedback"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserFeedbackListFeedbacks))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListFeedbacksByStatus List user feedbacks by status
func (c *UserFeedbackHTTPClientImpl) ListFeedbacksByStatus(ctx context.Context, in *ListFeedbacksByStatusRequest, opts ...http.CallOption) (*ListFeedbacksByStatusReply, error) {
	var out ListFeedbacksByStatusReply
	pattern := "/v1/userfeedback/status/{status}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserFeedbackListFeedbacksByStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateFeedback Update a user feedback
func (c *UserFeedbackHTTPClientImpl) UpdateFeedback(ctx context.Context, in *UpdateFeedbackRequest, opts ...http.CallOption) (*UpdateFeedbackReply, error) {
	var out UpdateFeedbackReply
	pattern := "/v1/userfeedback/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserFeedbackUpdateFeedback))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}