# backend/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/report/export/userfeedback:
        post:
            tags:
                - Report
            description: Export user feedback data as Excel
            operationId: Report_ExportUserFeedback
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/report.v1.ExportUserFeedbackRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/report.v1.ExportUserFeedbackReply'
    /v1/userfeedback:
        get:
            tags:
                - UserFeedback
            description: List user feedbacks with pagination
            operationId: UserFeedback_ListFeedbacks
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/userfeedback.v1.ListFeedbacksReply'
        post:
            tags:
                - UserFeedback
            description: Create a new user feedback
            operationId: UserFeedback_CreateFeedback
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/userfeedback.v1.CreateFeedbackRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/userfeedback.v1.CreateFeedbackReply'
    /v1/userfeedback/status/{status}:
        get:
            tags:
                - UserFeedback
            description: List user feedbacks by status
            operationId: UserFeedback_ListFeedbacksByStatus
            parameters:
                - name: status
                  in: path
                  required: true
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/userfeedback.v1.ListFeedbacksByStatusReply'
    /v1/userfeedback/{id}:
        get:
            tags:
                - UserFeedback
            description: Get a user feedback by ID
            operationId: UserFeedback_GetFeedback
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/userfeedback.v1.GetFeedbackReply'
        put:
            tags:
                - UserFeedback
            description: Update a user feedback
            operationId: UserFeedback_UpdateFeedback
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/userfeedback.v1.UpdateFeedbackRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/userfeedback.v1.UpdateFeedbackReply'
        delete:
            tags:
                - UserFeedback
            description: Delete a user feedback
            operationId: UserFeedback_DeleteFeedback
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/userfeedback.v1.DeleteFeedbackReply'
...
components:
    schemas:
...
        report.v1.ExportUserFeedbackReply:
            type: object
            properties:
                exportId:
                    type: string
                status:
                    type: string
                message:
                    type: string
            description: Reply for exporting user feedback
        report.v1.ExportUserFeedbackRequest:
            type: object
            properties:
                startDate:
                    type: string
                endDate:
                    type: string
                filters:
                    type: array
                    items:
                        type: string
                userId:
                    type: string
            description: Request for exporting user feedback
...
        userfeedback.v1.CreateFeedbackReply:
            type: object
            properties:
                feedback:
                    $ref: '#/components/schemas/userfeedback.v1.UserFeedbackData'
            description: Response messages
        userfeedback.v1.CreateFeedbackRequest:
            type: object
            properties:
                userId:
                    type: integer
                    format: uint32
                content:
                    type: string
                type:
                    type: string
                status:
                    type: string
                rating:
                    type: integer
                    format: int32
            description: Request messages
        userfeedback.v1.DeleteFeedbackReply:
            type: object
            properties:
                success:
                    type: boolean
        userfeedback.v1.GetFeedbackReply:
            type: object
            properties:
                feedback:
                    $ref: '#/components/schemas/userfeedback.v1.UserFeedbackData'
        userfeedback.v1.ListFeedbacksByStatusReply:
            type: object
            properties:
                feedbacks:
                    type: array
                    items:
                        $ref: '#/components/schemas/userfeedback.v1.UserFeedbackData'
                total:
                    type: string
        userfeedback.v1.ListFeedbacksReply:
            type: object
            properties:
                feedbacks:
                    type: array
                    items:
                        $ref: '#/components/schemas/userfeedback.v1.UserFeedbackData'
                total:
                    type: string
        userfeedback.v1.UpdateFeedbackReply:
            type: object
            properties:
                feedback:
                    $ref: '#/components/schemas/userfeedback.v1.UserFeedbackData'
        userfeedback.v1.UpdateFeedbackRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                userId:
                    type: integer
                    format: uint32
                content:
                    type: string
                type:
                    type: string
                status:
                    type: string
                rating:
                    type: integer
                    format: int32
        userfeedback.v1.UserFeedbackData:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                userId:
                    type: integer
                    format: uint32
                content:
                    type: string
                type:
                    type: string
                status:
                    type: string
                rating:
                    type: integer
                    format: int32
                createdAt:
                    type: string
                updatedAt:
                    type: string
            description: Data structure
...